refactor: LadderItem의 package를 domain -> view로 변경
